Extract Middleware and Properly Name noOp
